Create Online language learning platform
This Java program implements a basic user registration and login system for a language learning platform. It allows users to register an account with a username and password, and then log in using their credentials. The program utilizes basic concepts of object-oriented programming, such as classes, objects, and methods, as well as collections like HashMap for storing user data.